Fix SM120 DeepSeek V4 materialized decode

+++ vllm/model_executor/layers/deepseek_v4_attention.py
@@ -1181,7 +1181,7 @@ class DeepseekV4MLAAttention(nn.Module, AttentionLayerBase):
             ) = current_workspace_manager().get_simultaneous(
                 (
                     (num_decode_tokens, total_candidates, q.shape[-1]),
-                    torch.bfloat16,
+                    torch.float32,
                 ),
                 ((num_decode_tokens, total_candidates), torch.bool),
             )
